pyproject.toml
[project]
name = "agentgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.1",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
]
